--- modules/apps/headless/headless-web-experience/headless-web-experience-client/src/main/java/com/liferay/headless/web/experience/client/parser/v1_0/StructuredContentImageParser.java
@@ -1,50 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.headless.web.experience.client.parser.v1_0;
-
-import com.liferay.headless.web.experience.client.dto.v1_0.StructuredContentImage;
-
-import java.util.Collection;
-
-import javax.annotation.Generated;
-
-/**
- * @author Javier Gamarra
- * @generated
- */
-@Generated("")
-public class StructuredContentImageParser {
-
-	public static String toJSON(StructuredContentImage structuredContentImage) {
-		return null;
-	}
-
-	public static String toJSON(
-		Collection<StructuredContentImage> structuredContentImages) {
-
-		return null;
-	}
-
-	public static StructuredContentImage toStructuredContentImage(String json) {
-		return null;
-	}
-
-	public static StructuredContentImage[] toStructuredContentImages(
-		String json) {
-
-		return null;
-	}
-
-}
