+++ modules/apps/headless/headless-web-experience/headless-web-experience-client/src/main/java/com/liferay/headless/web/experience/client/serdes/v1_0/AggregateRatingSerDes.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.headless.web.experience.client.parser.v1_0;
+package com.liferay.headless.web.experience.client.serdes.v1_0;
 
 import com.liferay.headless.web.experience.client.dto.v1_0.AggregateRating;
 
@@ -25,7 +25,7 @@ import javax.annotation.Generated;
  * @generated
  */
 @Generated("")
-public class AggregateRatingParser {
+public class AggregateRatingSerDes {
 
 	public static String toJSON(AggregateRating aggregateRating) {
 		if (aggregateRating == null) {
